requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
gspread>=6.1.0
google-auth>=2.29.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
